fix compile error

+++ CS2_External/Utils/OffsetsUpdater.h
@@ -1,17 +1,16 @@
 #pragma once
+#include "cpp-httplib/httplib.h"
+#include "json/json.hpp"
 #include <iostream>
 #include <fstream>
 #include <map>
 #include <vector>
-#include <Windows.h>
-#include "cpp-httplib/httplib.h"
-#include "json/json.hpp"
 
 using json = nlohmann::json;
 
 namespace Updater
 {
-	static std::map<std::string, DWORD64> offsets;
+	static std::map<std::string, uintptr_t> offsets;
 
 	
 }
